add setting protection to patches
